requirements.txt
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
streamlit>=1.32.0,<2
pandas>=2.0.0,<4
pdfplumber>=0.11.0,<1
pytesseract>=0.3.10,<1
Pillow>=10.0.0
